requirements.txt
click==7.1.2
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==3.1
infinity==1.5
intervals==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.4
psycopg2-binary==2.8.6
six==1.15.0
validators==0.18.2
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Components==0.10.5
